refactor: remove subcomponents folder on product details component folder
