+++ src/pages/ProductDetails/ImageSlider.tsx
