+++ src/pages/ProductDetails/ProductDetails.tsx
@@ -2,8 +2,8 @@ import { useState } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
 import { Product } from '../../types';
 import useFetch from '../../hooks/useFetch';
-import Title from './subcomponents/Title';
-import ImageSlider from './subcomponents/ImageSlider';
+import Title from './Title';
+import ImageSlider from './ImageSlider';
 import QuantityInput from '../../components/QuantityInput/QuantityInput';
 import Loading from '../../components/Loading/Loading';
 import FeaturedProducts from '../../components/FeaturedProducts/FeaturedProducts';
